The server is now speaking in capital RSA and EC in crypto config key type.
I don't really remember this changing recently, but, the
AdminPortal did start displaying those key types in all
caps a while back. I didn't realize that was an ecosystem
wide change that would impact the representation of data.
Maybe this is not actually the correct fix to this, beacuse
everything references these key types.